# backend/docs/swagger/openapi.yml
openapi: 3.0.3
info:
  title: ERP Distribution - API Gateway
  version: 1.0.0
  description: |
    Specification OpenAPI centralisee exposee via l'API Gateway.
    Les clients consomment principalement ces routes /api/v1/*.
servers:
  - url: http://localhost:4000
    description: API Gateway locale
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: erp_session
paths:
...
  /api/v1/customers:
    get:
      summary: Lister les clients (via gateway)
      tags: [Customers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Liste clients
        '403':
          description: Permission customers.read requise
    post:
      summary: Creer un client (via gateway)
      tags: [Customers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id:
                  type: string
                  description: Optionnel (auto-genere si absent, ex CU-01234)
                customer_name:
                  type: string
                segment:
                  type: string
                city:
                  type: string
                region:
                  type: string
                email:
                  type: string
              required: [customer_name]
      responses:
        '201':
          description: Client cree
        '403':
          description: Permission customers.create requise
  /api/v1/customers/{customerId}:
    get:
      summary: Detail client (via gateway)
      tags: [Customers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail client
        '404':
          description: Non trouve
        '403':
          description: Permission customers.read requise
    patch:
      summary: Mettre a jour un client (via gateway)
      tags: [Customers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Client mis a jour
        '403':
          description: Permission customers.update requise
    delete:
      summary: Supprimer un client (via gateway)
      tags: [Customers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Client supprime
        '403':
          description: Permission customers.delete requise